creates routes module and moves auth into it
